Fix issue where console is visible if Cmder exists
When Cmder is running, the console window of Uncap remains visible. It
does not become hidden as expected.

The root cause of the issue here is that in the presence of Cmder, the
`FindWindow("ConsoleWindowClass", NULL)` call gets a handle to another
window (not the window in which Uncap is running), so the
`ShowWindow(h, SW_HIDE)` call later attempts to hide the wrong window
instead of the window in which Uncap is running.

This change fixes this issue by calling `GetConsoleWindow()` instead of
`FindWindow()`. The `GetConsoleWindow()` call gets a handle to the
current window regardless of whether Cmder is running or not.

Further, this change undoes the earlier change in
commit d81d1066d07aff47f833a7f6a7070dba947e2e51 which fixed this issue
with a workaround. That workaround is no longer necessary because we
address the root cause of the issue in this change.