Support mapping of injected keyboard events
Prior to this change, injected keyboard events were being ignored by
Uncap. This was done in order to avoid remap loops. A remap loop may
occur when two keys are mapped to each other. For example, if `A` is
mapped to `B` and vice versa, then on pressing `A`, Uncap receives a
keyboard event for `A` and it replaces it with an injected keyboard
event for `B`. The injected keyboard event for `B` is again received by
Uncap but it ignores this injected keyboard event to avoid remapping it
to `A` again which would cause a remap loop.

Prior to this change, Uncap avoided such remap loops by ignoring
keyboard events which had its `LLKHF_INJECTED` flag set. But this meant
that genuine keyboard events injected by other tools such as On-Screen
Keyboard were also ignored by Uncap. Thus Uncap failed to map a key that
was pressed on such tools.

With this change, Uncap does not ignore injected keyboard events
anymore. In order to avoid remap loop, it sets the `dwExtraInfo` member
of the injected keyboard event with a magic value and ignores any
received keyboard event with the same magic value.